Fix sounds compilation problems
Stopped copying over peach's lines from the mario folder to wario/luigi sound folders
Stopped compiling peach's lines for wario/luigi sound folders
Created a tool to precompile the aiffc tables and append them into custom sounds
    this matches what mario's sound files are like

I believe the issue is that previously we would copy over the sound files and
they were either incomplete or would be compiled during the copying process

I noticed that Mario's sounds had their sound tables inserted into the aiff
file already, and none of our custom sounds did. This caused a different path
to be taken for mario sounds than the rest of ours. In order to prevent the
possibility that there was a problem with that seperate path, I've created
a tool that can insert the table into the file to match how mario sounds are.